Dilation Models

Dilation Overview

Dilation models describe how the polymer volume changes as a function of pressure. They all have the following shared functionality.

SorptionModels.fit_model — Method

fit_model(model, pressures_mpa, fractional_dilations, [uncertainty_method])

Fit a dilation model to a set of pressures and fractional dilations.

Any keyword arguments get passed on to the specific model.

uncertainty_method can be any supported method of propagating uncertainty within the model.

SorptionModels.predict_dilation — Method

predict_dilation(model::SorptionModel, args...)

Predict fractional dilation with a dilation model. Predictions are generally returned in the format of a vector of values corresponding to an input of a vector of pressures.

SorptionModels.predict_dilation_derivative — Method

Get the derivative of a dilation vs pressure curve at a given pressure.

Dual Mode Dilation

SorptionModels.fit_model — Function

fit_model(::DualModeDilation, pressures_mpa, frac_dilations, uncertainty_method=nothing; kwargs...)

Fit the dilation data to an dual mode dilation model, using a dual mode model previoulsy fit to concentration and pressure data.

Dilation is described by two additional parameters, $V_D$ and $f$ using the following model:

${\Delta}{V} / V_0 = V_D (k_D \cdot p + \frac{f C_{H}^{'}bp}{1+bp})$

$V_D$ represents the effective condensed penetrant molar volume.

$f$ represents an empirical fit describing how much the langmuir mode participates in dilation.

J.D. Moon, M. Galizia, H. Borjigin, R. Liu, J.S. Riffle, B.D. Freeman, D.R. Paul, Water Vapor Sorption, Diffusion, and Dilation in Polybenzimidazoles, Macromolecules. 51 (2018) 7197–7208. https://doi.org/10.1021/acs.macromol.8b01659.

Empirical Dilation

SorptionModels.fit_model — Function

fit_model(::EmpiricalDilation, pressures_mpa, frac_dilations, uncertainty_method=nothing; n_params=3, kwargs...)

Fit the dilation data to an empirical dilation model with no physical meaning. This model is intended to fit any dilation curve given enough parameters. The user may specify up to 4 empirical paramteres. Loosely, the use of each parameter is as follows:

n_params = 1: Linear dilation

n_params = 2: Naive dual mode like dilation

n_params = 3: Dual mode like dilation

n_params = 4: Dual mode like dilation with concave down or tapering behavior permitted

If using uncertainty_method (:JackKnife or :Hessian implemented), each extra parameter will greatly increase the resulting uncertainty, as this model is completely empirical.