Writer Methods

SorptionModels.write_analysis — Method

write_analysis(analysis::DualModeDesorption, workbook::XLSX.XLSXFile; [name])
write_analysis(analysis::DualModeDesorption, filepath::AbstractString; [name])

Write a Dual Mode Desorption analysis out to a .xlsx workbookor an .xlsxfilepath. If a worksheetname` is not provided, default names are created.

See DualModeDesorption

SorptionModels.write_analysis — Method

write_analysis(analysis::AbstractIsostericHeatAnalysis, workbook::XLSX.XLSXFile; [name])
write_analysis(analysis::AbstractIsostericHeatAnalysis, filepath::AbstractString; [name])

Write an isosteric heat analysis out to a .xlsx workbookor an .xlsxfilepath. If a worksheetnameis not provided, default names are created. See [IsostericHeatAnalysis`](@ref)

SorptionModels.write_analysis — Method

write_analysis(analysis::MolarVolumeAnalysis, workbook::XLSX.XLSXFile; [name])
write_analysis(analysis::MolarVolumeAnalysis, filepath::AbstractString; [name])

Write a molar volume analysis out to a .xlsx workbookor an .xlsxfilepath. If a worksheetname` is not provided, default names are created.

See MolarVolumeAnalysis

SorptionModels.write_analysis — Method

write_analysis(analysis::PartialImmobilizationModel, workbook::XLSX.XLSXFile; [name])
write_analysis(analysis::PartialImmobilizationModel, filepath::AbstractString; [name])

Write a partial immobilization analysis out to a .xlsx workbookor an .xlsxfilepath. If a worksheetname` is not provided, default names are created.

See PartialImmobilizationModel

SorptionModels.write_analysis — Method

write_analysis(analysis::VantHoffDualModeAnalysis, workbook::XLSX.XLSXFile; [name])
write_analysis(analysis::VantHoffDualModeAnalysis, filepath::AbstractString; [name])

Write a Vant Hoff Dual Mode analysis out to a .xlsx workbookor an .xlsxfilepath. If a worksheetname` is not provided, default names are created.

See VantHoffDualModeAnalysis

SorptionModels.write_analysis — Method

write_analysis(analysis::ZimmLundbergAnalysis, workbook::XLSX.XLSXFile; [name])
write_analysis(analysis::ZimmLundbergAnalysis, filepath::AbstractString; [name])

Write a Zimm Lundberg analysis out to a .xlsx workbookor an .xlsxfilepath. If a worksheetname` is not provided, default names are created.

See ZimmLundbergAnalysis

Mobility and Thermodynamic factors do not have writers (yet!)